style: 매장 상품 검색 스타일 수정
