+++ src/App.tsx
@@ -1,7 +1,7 @@
 import { Global, ThemeProvider } from '@emotion/react';
 import theme from '@styles/theme';
 import GlobalStyle from './styles/global';
-import StoreSearchBar from '@components/StoreSearch/StoreSearchBar';
+import StoreSearchBar from './components/StoreSearchBar/StoreSearchBar';
 
 function App() {
   return (
